prechadzame do testovacieho rezimu #66

+++ src/generate_scripts/clear_time_slot.php
@@ -3,23 +3,16 @@ include('../db.php');
 include('dump.php');
 date_default_timezone_set("Europe/Bratislava");
 
-$next_start_point_of_generation = strtotime('2 week ago'); //  treba specifikovat format generovania napr. UTC 00:00
+$next_start_point_of_generation = strtotime('1 week ago'); //  1 tizdne koli tomu ako sme sa dohodli s p.p
 $date = date("Y-m-d", $next_start_point_of_generation);
 echo 'DATA cistime k datumu  '. $date . ' <br> ';
-$sql = "delete from time_slot where TIMESTAMP ($date) < end_date_time  ";
-if ($result = $mysqli->query($sql)) {  // vykonaj dopyt
-    echo "VYCISTENA DB(TIME SLOTS)<br>";
-} else{
-    echo "CHYBA SKRIPTU <br> ";
+if (!$mysqli->connect_errno) {
+    $sql = "delete from time_slot where TIMESTAMP ($date) < end_date_time  ";
+    if ($result = $mysqli->query($sql)) {  // vykonaj dopyt
+        echo "VYCISTENA DB(TIME SLOTS)<br>";
+    } else{
+        echo "CHYBA SKRIPTU <br> ";
+    }
+}else {
+    echo 'Serverova chyba databaza nieje pripojena';
 }
-
-// Pokial to chces proste cele premzat len to odkomentuj !!!!
-/*
-
-$sql = "delete from time_slot where id > 0   ";
-if ($result = $mysqli->query($sql)) {  // vykonaj dopyt
-    echo "VYCISTENA DB(TIME SLOTS)<br>";
-} else{
-    echo "CHYBA SKRIPTU <br> ";
-}
-*/
